Fork style attempt (#9)
* Utilize flex, style input

By utilizing flexbox we can reduce the amount of css needed to layout the page.

Some sort of attempt at styling the interface input.
Removes the generic `button` in favour of more specific styling. Doing this allows us to
experiement more with styles. We can create more generic components as we need them.

* chore: use npm

* layout playlist song selection component

* test search route

* feat: Update some folder structure, style updates